+++ components/esm3/creaturelevliststate.cpp
@@ -12,7 +12,7 @@ namespace ESM
 
         if (esm.getFormatVersion() <= MaxActorIdSaveGameFormatVersion)
         {
-            mSpawnedActor.mIndex = -1;
+            mSpawnedActor.mIndex = static_cast<uint32_t>(-1);
             esm.getHNOT(mSpawnedActor.mIndex, "SPAW");
         }
         else if (esm.peekNextSub("SPAW"))
